@@ -71,7 +71,7 @@ class FormsController < ApplicationController
       where('events.master_id IS NULL')
     @pages = Page.editable(current_user)
     @cancel_path = session[:edit_form_cancel_path] || page_path(@page)
-    @possible_parents = @form.page.forms.delete_if{|f| f.id == @form.id}
+    @possible_parents = @form.page.forms.to_a.delete_if{|f| f.id == @form.id}
   end
   
   def edit_fields
@@ -158,7 +158,8 @@ class FormsController < ApplicationController
     params.require(:form).permit(:name, :page_id, :event_id,
       :payable, :published, :pay_by_check, :pay_by_paypal,
       :updated_by, :version, :parent_id, :authenticated,
-      :many_per_user, :authentication_text).merge(:updated_by => current_user)
+      :many_per_user, :authentication_text,
+      :call_to_action).merge(:updated_by => current_user)
   end
   
 end
