@@ -3,8 +3,8 @@ class Conversation < ActiveRecord::Base
   belongs_to :user
   has_many :comments, -> { order('created_at ASC') }
   
-  validate :text, :presence => true
-  validate :user, :presence => true
+  validates :text, :presence => true
+  validates :user, :presence => true
   
   def authorized?(user)
     page.authorized?(user)
