@@ -2,9 +2,9 @@ require File.expand_path('../boot', __FILE__)
 
 require 'rails/all'
 
-# If you have a Gemfile, require the gems listed there, including any gems
+# Require the gems listed in Gemfile, including any gems
 # you've limited to :test, :development, or :production.
-#Bundler.require(:default, Rails.env)
+#Bundler.require(*Rails.groups)
 Bundler.require *Rails.groups(:assets => %w(development test))
 
 module Pbc3
@@ -13,16 +13,6 @@ module Pbc3
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
 
-    # Custom directories with classes and modules you want to be autoloadable.
-    # config.autoload_paths += %W(#{config.root}/extras)
-
-    # Only load the plugins named here, in the order given (default is alphabetical).
-    # :all can be used as a placeholder for all plugins not explicitly named.
-    # config.plugins = [ :exception_notification, :ssl_requirement, :all ]
-
-    # Activate observers that should always be running.
-    # config.active_record.observers = :cacher, :garbage_collector, :forum_observer
-
     # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
     # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
     # config.time_zone = 'Central Time (US & Canada)'
@@ -30,19 +20,15 @@ module Pbc3
     # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
     # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
     # config.i18n.default_locale = :de
-
-    # JavaScript files you want as :defaults (application.js is always included).
-
-    # Configure the default encoding used in templates for Ruby 1.9.
-    config.encoding = "utf-8"
+    
+    # Do not swallow errors in after_commit/after_rollback callbacks.
+    config.active_record.raise_in_transactional_callbacks = true
 
     # Configure sensitive parameters which will be filtered from the log file.
-    config.filter_parameters += [:password]
+    ##config.filter_parameters += [:password]
     
-    #config.assets.paths << "#{Rails.root}/app/assets/stylesheets/mobile/mobile.css.scss"
-    #config.assets.paths << "#{Rails.root}/app/assets/javascripts/mobile/mobile.js.coffee"
-    config.assets.precompile += ["mobile/mobile.css", "mobile/mobile.js"]
+    ##config.assets.precompile += ["mobile/mobile.css", "mobile/mobile.js"]
     
-    config.i18n.enforce_available_locales = false
+    ##config.i18n.enforce_available_locales = false
   end
 end
