@@ -1,27 +1,25 @@
-Pbc3::Application.configure do
-  # Settings specified here will take precedence over those in config/application.rb
-  
+Rails.application.configure do
+  # Settings specified here will take precedence over those in config/application.rb.
+
   # In the development environment your application's code is reloaded on
-  # every request.  This slows down response time but is perfect for development
-  # since you don't have to restart the webserver when you make code changes.
+  # every request. This slows down response time but is perfect for development
+  # since you don't have to restart the web server when you make code changes.
   config.cache_classes = false
 
   # Do not eager load code on boot.
   config.eager_load = false
 
-  # Show full error reports and disable caching
+  # Show full error reports and disable caching.
   config.consider_all_requests_local       = true
   config.action_controller.perform_caching = false
 
-  # Don't care if the mailer can't send
-  config.action_mailer.delivery_method = :sendmail
-  config.action_mailer.raise_delivery_errors = true
-  config.action_mailer.asset_host = "http://localhost"
+  # Don't care if the mailer can't send.
+  config.action_mailer.raise_delivery_errors = false
 
-  # Print deprecation notices to the Rails logger
+  # Print deprecation notices to the Rails logger.
   config.active_support.deprecation = :log
-  
-  # Raise an error on page load if there are pending migrations
+
+  # Raise an error on page load if there are pending migrations.
   config.active_record.migration_error = :page_load
 
   # Debug mode disables concatenation and preprocessing of assets.
@@ -29,15 +27,23 @@ Pbc3::Application.configure do
   # number of complex assets.
   config.assets.debug = true
 
-  # Only use best-standards-support built into browsers
-  ###config.action_dispatch.best_standards_support = :builtin
+  # Asset digests allow you to set far-future HTTP expiration dates on all assets,
+  # yet still be able to expire them through the digest params.
+  config.assets.digest = true
+
+  # Adds additional error checking when serving assets at runtime.
+  # Checks for improperly declared sprockets dependencies.
+  # Raises helpful error messages.
+  config.assets.raise_runtime_errors = true
+
+  # Raises error for missing translations
+  # config.action_view.raise_on_missing_translations = true
   
   config.action_mailer.default_url_options = { :host => 'UNCONFIGURED' }
   
   config.time_zone = "Pacific Time (US & Canada)"
   
   config.after_initialize do
-    Configuration.mailman_dir = '/usr/local/bin'
     Configuration.paypal_url = 'https://www.sandbox.paypal.com/cgi-bin/webscr'
   end
 end
