@@ -1,4 +1,6 @@
-Pbc3::Application.routes.draw do
+Rails.application.routes.draw do
+  # The priority is based upon order of creation: first created -> highest priority.
+  # See how all your routes lay out with "rake routes".
 
   get "hyper_home/index"
 
