@@ -11,7 +11,10 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20140524180403) do
+ActiveRecord::Schema.define(version: 20141202130301) do
+
+  # These are extensions that must be enabled in order to support this database
+  enable_extension "plpgsql"
 
   create_table "audios", force: true do |t|
     t.string   "caption"
@@ -221,6 +224,7 @@ ActiveRecord::Schema.define(version: 20140524180403) do
     t.boolean  "authenticated",       default: false
     t.boolean  "many_per_user",       default: false
     t.text     "authentication_text"
+    t.string   "call_to_action"
   end
 
   create_table "holidays", force: true do |t|
@@ -489,14 +493,15 @@ ActiveRecord::Schema.define(version: 20140524180403) do
     t.string   "bio_back_content_type"
     t.integer  "bio_back_file_size"
     t.datetime "bio_back_updated_at"
-    t.integer  "bio_back_color",             default: 0
-    t.integer  "banner_text_color",          default: 0
+    t.integer  "bio_back_color",               default: 0
+    t.integer  "banner_text_color",            default: 0
     t.string   "child_feature_file_name"
     t.string   "child_feature_content_type"
     t.integer  "child_feature_file_size"
     t.datetime "child_feature_updated_at"
-    t.integer  "child_feature_text_color",   default: 0
+    t.integer  "child_feature_text_color",     default: 0
     t.integer  "updated_by"
+    t.boolean  "hero_text_background_overlay", default: true
   end
 
   create_table "users", force: true do |t|
